tools/data_validator.py
from anthropic.mcp import MCPClient  # Hypothetical 2025 SDK

mcp_client = MCPClient()


@mcp_client.tool
def validate_data(data: str) -> dict:
    # MCP-handled validation
    return {"valid": True, "issues": []}
