pyproject.toml
[project]
name = "multi-agent-data-engineering-swarm"
version = "0.1.0"
description = "A multi-agent swarm system for automated data engineering pipelines using LangChain and various LLM providers"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.0.15",
    "faiss-cpu>=1.11.0.post1",
    "langchain>=0.3.26",
    "langchain-anthropic>=0.3.17",
    "langchain-deepseek>=0.1.3",
    "langchain-ollama>=0.3.5",
    "langchain-openai>=0.3.28",
    "langchain-xai>=0.2.4",
    "langgraph>=0.5.4",
    "pandas>=2.3.1",
    "pydantic>=2.11.7",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
]
